Quite a big commit for the  html - Edited the loginpage name to sign in instead and also created a sign up page but not much is on it.
Cleaned up the main page with some more comments and some easier/simpler class names additionally, more functionality for the search bar has been added.

Created the html for the account page with comments. Not much is clear on what to add to the page but very basic functionality has been added onto it.

If theres something missing let me know please and ill get on it. Thanks