**IMPORTANT** I have moved the HTML and JavaScript files into a new folder named "main" please keep it like this.
Major Additions:
I have created a nodejs server that will be responsible for http requests, for example when the admin tries to create a new recipe that data will be fetched into a JSON file format. which will then be stored within the main folder

**HOW TO RUN**
In the terminal type "node server.js" to run the server and you should get a prompt to open in browser which you do, the website runs as normal and begins at the index page where I have added a link to the admin page, click here and use the 3 input boxes to add the information of the new recipe. Finally click on the add recipe button and once the alert has been accepted the new JSON file should appear in the folder.