Additions to the account page
Added some HTML to display the fullname, email and phone number for a users account.

Not sure if the fullname or other displays work until I get an example account or if theres is a function so it is untested.

Also phone number isnt included in the sign up process, will add if the phone number is necessary or wont break anything.