I have updated the server side and js to accomodate for a new approach as well as added an ingredients class to have multiple ingredients added to the recipe rather than just one string of ingredients
The ingredients class is a super class with 3 subclasses that will accomodate for the different types of ingredients that will be inputted. This was achieved by distinguishing between the different measurement types required for different ingredients then creating their respective class. Furthermore the javascript code was updated by initiating an ingredientsarray that will hold the new ingredients from the list in the html page. The rest of the functionality is the same however changes were made to make a connection to our database where the recipes are stored instead of writing to a json file.

the js now establishes a connection to our SQLDatabase with doing so the server.js has been updated by instead of handling json file creation it instead handles SQL queries as well as the ingredient formatting required for the database. Error handling was also implemented in case of faulty connection with the database