HTML and CSS for checkboxes done
Played with the html layout so clicking the actual sentences of the checkboxes would tick them for easier user usability.

Implemented a hover and pop animation which could be used on other areas of our app.

Made use of "type" CSS so other checkboxes we may add in the future can apply the same styling.